Created new html view, changed project structure, removed index.js, created basic api requests
