+++ src/views/index.html
@@ -4,7 +4,7 @@
   <head>
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Node.js Starter</title>
+    <title>Tune Table</title>
     <link rel="stylesheet" type="text/css" href="/css/styles.css" />
   </head>
   <body>
@@ -22,6 +22,8 @@
 
     <h1>Hello!</h1>
     <p>If you can see this text, your server is running properly!</p>
-    <p>This is for github example, very cool. Another change</p>
+    
+    <button type="button">Display</button>
+
   </body>
 </html>
