+++ src/views/users.html
@@ -1,0 +1,15 @@
+<!--views/index.html-->
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Node.js Starter</title>
+    <link rel="stylesheet" type="text/css" href="/css/styles.css" />
+  </head>
+  <body>
+
+    <button type="button">Display</button>
+    
+  </body>
+</html>
